Changed the dimension of et from int to float

+++ rails/wx-services/app/apis/period_struct.rb
@@ -33,5 +33,5 @@ class PeriodStruct < ActionWebService::Struct
   member :uv, :int
   member :high_uv, :int
 
-  member :et, :int
+  member :et, :float
 end
